+++ packages/shadenfreude/src/compilers.ts
@@ -1,0 +1,9 @@
+import { Variable } from "./variables"
+
+export const compileShader = (root: Variable) => {
+  const vertexShader = ``
+
+  const fragmenShader = ``
+
+  return { vertexShader, fragmenShader }
+}
