+++ packages/shadenfreude/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./lib/concatenator3000"
-export * from "./shadenfreude"
+export * from "./variables"
+export { compileShader } from "./compilers"
